feat(voice) Architus can record and send raw WAV file
Adds start_recording and stop_recording commands to architus via the
record_cog. start_recording will start architus listening the the
commanding user's voice channel. stop_recording will stop the bot
listening to the channel and send a WAV file of the audio to
the text channel the command was sent from.