feat(auto response) Improves ux for removing auto responses
User can now copy and paste the full `trigger::response` text and
pass that to the `!remove` command to remove the auto response
instead of having to parse out just the trigger from the auto
response and pass that to the `!remove` command.

In addition, when the speech bubble react is used to see where
an auto response comes from, any user with remove permissions
can react with an :x: to remove that auto response.