Add submission & event store components for logs feature (#157)
## Description

This PR takes out the submission (logs/submission) and event store (elasticsearch) components from #100 to move them into a smaller PR that can be reviewed and deployed to staging separately. Additionally, it includes the logs common lib (lib/logs-lib-rs) that can be used to submit events to the event store via the submission service from Rust in an ergonomic way. See #100 for more context on the changes.